Refactored to add Application class
