+++ build.gradle
@@ -8,7 +8,7 @@ plugins {
 
 group 'org.radarcns'
 version '0.5.0-SNAPSHOT'
-mainClassName = 'org.radarcns.hdfs.RestructureAvroRecords'
+mainClassName = 'org.radarcns.hdfs.Application'
 
 sourceCompatibility = '1.8'
 targetCompatibility = '1.8'
